fix for some directives
- added preg_quote() to support directives with "."
- added support any count of spaces before and after "=" in directive row

+++ assistant.php
@@ -8,28 +8,30 @@ if(!is_array($arConfig) && count($arConfig) < 1) die('Bad config');
 $arPresets = [];
 
 // Parse presets
+$arReplacements = [];
 foreach($arConfig['presets'] as $arPreset){
     foreach($arPreset['directives'] as $directiveName => $directive){
-        $patterns[$arPreset['name']][] = '#^;?' . $directiveName . ' = (.*)#m';
+        $directiveNameRegex = preg_quote($directiveName);
+        $patterns[$arPreset['name']][] = '#^;?' . $directiveNameRegex . '\s*=\s*?(.*)#m';
         if(isset($directive['disabled']) && $directive['disabled'] == true){
-            $raplacements[$arPreset['name']][] = '; ' . $directiveName .' = ' . $directive['value'];
+            $arReplacements[$arPreset['name']][] = '; ' . $directiveName .' = ' . $directive['value'];
         }else{
-            $raplacements[$arPreset['name']][] = $directiveName .' = ' . $directive['value'];
+            $arReplacements[$arPreset['name']][] = $directiveName .' = ' . $directive['value'];
         }
     }
 }
 
 // Write configs
 foreach($arConfig['files'] as $arFileConfig){
-    if(is_file($arFileConfig['path']) && $arFileConfig['preset'] != ''){
+    if(is_file($arFileConfig['path']) && $arFileConfig['preset'] != '' && $arReplacements[$arFileConfig['preset']]){
         echo 'Write config preset "' . $arFileConfig['preset'] . '" for: ' . $arFileConfig['path'] . PHP_EOL;
         if(!is_file($arFileConfig['path'].'.orig')){
             copy($arFileConfig['path'], $arFileConfig['path'].'.orig');
         }
         copy($arFileConfig['path'], $arFileConfig['path'].'.saved');
-    
+
         $srcConfig = file_get_contents($arFileConfig['path']);
-        $resConfig = preg_replace($patterns[$arFileConfig['preset']], $raplacements[$arFileConfig['preset']], $srcConfig);
+        $resConfig = preg_replace($patterns[$arFileConfig['preset']], $arReplacements[$arFileConfig['preset']], $srcConfig);
         file_put_contents($arFileConfig['path'], $resConfig);
     }
 }
